refactor(web): move ai from shared to zero
