+++ turbo/apps/web/src/lib/zero/ai/__tests__/lightweight-model.test.ts
