+++ turbo/apps/web/src/lib/zero/ai/lightweight-model.ts
